+++ SakuraTree.pro
@@ -68,9 +68,9 @@ SOURCES += \
     src/branch_builder/provision_branch_builder.cpp \
     src/processing/blossoms/ssh/ssh_scp_blossom.cpp \
     src/processing/blossoms/ssh/ssh_cmd_blossom.cpp \
-    src/processing/blossoms/sakura/sakura_copy_subtree_blossom.cpp \
+    src/processing/blossoms/special/sakura_copy_subtree_blossom.cpp \
     src/processing/blossoms/ssh/ssh_cmd_create_file_blossom.cpp \
-    src/processing/blossoms/sakura/print_blossom.cpp \
+    src/processing/blossoms/special/print_blossom.cpp \
     src/processing/process_execution.cpp \
     src/processing/blossoms/files/file_methods.cpp \
     src/processing/blossoms/files/common_files/file_copy_blossom.cpp \
@@ -85,7 +85,7 @@ SOURCES += \
     tests/processing/blossoms/files/common_files/file_copy_blossom_test.cpp \
     tests/processing/blossoms/files/common_files/file_delete_blossom_test.cpp \
     tests/processing/blossoms/files/common_files/file_rename_blossom_test.cpp \
-    src/processing/blossoms/sakura/template_blossom.cpp
+    src/processing/blossoms/template/template_create_blossom.cpp
 
 
 HEADERS += \
@@ -111,10 +111,10 @@ HEADERS += \
     src/branch_builder/provision_branch_builder.h \
     src/processing/blossoms/ssh/ssh_scp_blossom.h \
     src/processing/blossoms/ssh/ssh_cmd_blossom.h \
-    src/processing/blossoms/sakura/sakura_copy_subtree_blossom.h \
+    src/processing/blossoms/special/sakura_copy_subtree_blossom.h \
     src/processing/blossoms/ssh/ssh_cmd_create_file_blossom.h \
     src/processing/sakura_tree_callbacks.h \
-    src/processing/blossoms/sakura/print_blossom.h \
+    src/processing/blossoms/special/print_blossom.h \
     src/processing/process_execution.h \
     src/processing/blossoms/files/file_methods.h \
     src/processing/blossoms/files/common_files/file_copy_blossom.h \
@@ -129,6 +129,6 @@ HEADERS += \
     tests/processing/blossoms/files/common_files/file_copy_blossom_test.h \
     tests/processing/blossoms/files/common_files/file_delete_blossom_test.h \
     tests/processing/blossoms/files/common_files/file_rename_blossom_test.h \
-    src/processing/blossoms/sakura/template_blossom.h
+    src/processing/blossoms/template/template_create_blossom.h
